% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clcurve.R
\name{cl_curve_plot}
\alias{cl_curve_plot}
\title{Plot clustered ICE curves using ggplot2}
\usage{
cl_curve_plot(
  data = NULL,
  icevalue,
  gg,
  yhat,
  gr = NULL,
  sz = 0.5,
  aa = 1/100,
  xlab = NULL,
  ylab = NULL,
  fct = TRUE,
  xvalue = NULL
)
}
\arguments{
\item{data}{data.frame with the data to be ploted}

\item{icevalue}{Name of ice curve variable.}

\item{gg}{Name of the group id variable.}

\item{yhat}{Name of each ice curve id.}

\item{gr}{vector with the selected groups to be ploted.}

\item{sz}{line widht in the plot.}

\item{aa}{numeric value for transparency in the plot.}

\item{xlab}{string with x axis label.}

\item{ylab}{string with y axis label.}

\item{fct}{logical value to inicate facets in the plot.}

\item{xvalue}{variable name with grid values ploted in x axis.}
}
\value{
A ggplot2 object
}
\description{
Plot clustered ICE curves using ggplot2
}
\examples{
\dontrun{
cl_curve_plot(data = curveDT, gg = grlab, gr = 1:4, xvalue = grid.val, aa =1/20)
}
}
