% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordDT.R
\docType{data}
\name{coordDT}
\alias{coordDT}
\title{Data with location information asociated to ice curves values}
\format{
A data frame with 214 rows and 10 variables
}
\usage{
data(coordDT)
}
\description{
data.frame with 1000 rows and five columns
\itemize{
\item{long}{ Group id}
\item{lat}{ refractive index}
\item{grlab}{ Grid id}
\item{yhat.id}{ Ice curve values}
}
}
\keyword{datasets}
