% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveDT.R
\docType{data}
\name{curveDT}
\alias{curveDT}
\title{Data with ICE curve information}
\format{
A data frame with 214 rows and 10 variables
}
\usage{
data(curveDT)
}
\description{
data.frame with 1000 rows and five columns
\itemize{
\item{grlab}{ Group id}
\item{yhat.id}{ ice curve id}
\item{grid.point}{ Grid id}
\item{ice}{ Ice curve values}
\item{gridval}{ Grid value}
}
}
\keyword{datasets}
