% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_string_ppi_matrix.R
\name{get_string_ppi_matrix}
\alias{get_string_ppi_matrix}
\title{Protein-protein interaction (PPIs) from STRING.}
\usage{
get_string_ppi_matrix(
  genes,
  directed = FALSE,
  version = "11",
  species = 9606,
  score.threshold = 0,
  string.dir = "",
  max.homology.bitscore = Inf,
  benchmark.pathway = NULL,
  normalize.score = T
)
}
\arguments{
\item{genes}{character vector. gene names (HGNC gene symbols).}

\item{directed}{logical. Directed interactions?}

\item{version}{character. Version of STRING database.}

\item{species}{numeric. Species code for STRING (9606 for homo sapiens).}

\item{score.threshold}{numeric. A value between 0 and 1000.
Any score below \code{score.threshold} will be replaced by 0.}

\item{string.dir}{character. Local directory where STRING database
will be stored locally so that it can be used offline.
If \code{string.dir = ""}, a temporary directory is used.}

\item{max.homology.bitscore}{numeric. Maximum homology score.
Use \code{0} to filter all interactions between homologous proteins, and
\code{Inf} not to filter any interaction.}

\item{benchmark.pathway}{character. Pathway type to benchmark.
Examples: \code{"KEGG", "REACTOME", "BIOCARTA", "Disease",
"Pfam", "NCI", "ECOCYC"}.
Use \code{NULL} to avoid benchmarking.
If a pathway type is given, an interaction will have a non-zero score
only if both genes are present in at least one pathway.}

\item{normalize.score}{logical. Should the scores be normalized?
If \code{TRUE}, the scores will be divided by 1000.
Consequently, each score will be between 0 and 1.}
}
\value{
A 2-dimensional matrix with interactions between the given genes.
Note: any gene that cannot be mapped to a protein in STRING will be excluded.
}
\description{
This function generates a matrix of PPI scores available in STRING database
within a given set of genes.
}
\details{
Note: all the genes are expected as HGNC gene symbols.
}
\examples{
genes = c("TP53", "RBM3", "SF3", "LIM12", "MDM4", "TMEM160", "TP53BP2", "MDM2", "PDR", "MEG3", "EGFR")
interactions = get_string_ppi_matrix(genes, version = "11")
}
