% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_delimitted_string.R
\name{parse_delimitted_string}
\alias{parse_delimitted_string}
\title{Function to parse delimitted string}
\usage{
parse_delimitted_string(x, delim = ",", rm.empty = T)
}
\arguments{
\item{x}{character. String to be parsed.}

\item{delim}{character. The delimiter.}

\item{rm.empty}{logical. If TRUE, removes every empty string after parsing.}
}
\value{
A vector of parsed character strings.
}
\description{
This function splits the given string by a delimiter.
}
\examples{
x = "one,two,three,"
parsed_strings = parse_delimitted_string(x, delim = ',', rm.empty = T)
}
