% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Function to compute and visualize optimum alpha cluster values}
\usage{
plot_alpha(
  DD1,
  DD0,
  resp = NULL,
  nfrac = 0.2,
  yhat,
  icevalue,
  grid,
  clRange = 3:5,
  alpRange = seq(0, 1, 0.2)
)
}
\arguments{
\item{DD1}{data.frame with coordinate information asociated to ice curves values named lat and long.}

\item{DD0}{data.frame with ICE curve information.}

\item{resp}{response vector to perform stratified sample}

\item{nfrac}{sampling fraction}

\item{yhat}{Name of each ice curve id}

\item{icevalue}{Name of ice curve variable.}

\item{grid}{Name of the grid value variable.}

\item{clRange}{Numeric vector with posible values of number of cluster.}

\item{alpRange}{Numeric vector with posible values of alpha parameter.}
}
\value{
A list with cluster metric for each value of alpRange and clRange, and a visualization of the intertia decomposition.
}
\description{
Function to compute and visualize optimum alpha cluster values
}
\examples{
\dontrun{
plot_alpha( DD0 = curveDT, DD1 = coordDT, yhat = yhat.id, icevalue = ice, grid = grid.val)
}
}
