% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clcoord.R
\name{plot_clcoord}
\alias{plot_clcoord}
\title{Plot map including geographical location of observations colored by groups using ggmap}
\usage{
plot_clcoord(
  data = NULL,
  gg,
  gr = NULL,
  sz = 0.5,
  aa = 0.7,
  fct = TRUE,
  zz = 13,
  long,
  lat,
  region = NULL
)
}
\arguments{
\item{data}{data.frame with the data to be ploted having the variables longitude and latitude, named long and lat}

\item{gg}{name of the group id variable.}

\item{gr}{vector with the selected groups to be ploted.}

\item{sz}{line widht in the plot.}

\item{aa}{numeric value for transparency in the plot.}

\item{fct}{logical value to inicate facets in the plot.}

\item{zz}{a zoom level for map}

\item{long}{variable name with the longitude data.}

\item{lat}{variable name with latitude data.}

\item{region}{a bounding box in the format c(lowerleftlon, lowerleftlat, upperrightlon, upperrightlat), see example}
}
\value{
A ggplot2 object.
}
\description{
Plot map including geographical location of observations colored by groups using ggmap
}
\examples{
\dontrun{
montevideo <- c(left = -56.286532, bottom = -34.95, right = -56.004532, top =-34.801112 )
plot_clcoord(data = coordDT, gg = grlab, gr = 1:4, region = montevideo)
}
}
