% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spice.R
\name{spice}
\alias{spice}
\title{SPICE: Spanning tree based inference of co-expression networks}
\usage{
spice(
  expr,
  method = "pearson",
  iter = 100,
  frac.row = 1,
  frac.col = 0.8,
  n.cores = 1,
  rank.ties = "average",
  filter.mat = NULL,
  weight.method = "qnorm",
  adjust.weight = T,
  powerVector = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  RsquaredCut = 0.8,
  removeFirst = FALSE,
  nBreaks = 10,
  adjust.clr = F,
  seed = NULL,
  verbose = F
)
}
\arguments{
\item{expr}{data.frame or matrix. Processed gene expression data (gene x sample). See details.}

\item{method}{a character string indicating how to compute association between a pair of genes.
This must be one of "pearson" (default), "spearman", "mi.empirical", "mi.spearman", and "mi.pearson". See details.}

\item{iter}{numeric. Number of iterations.}

\item{frac.row}{numeric. Fraction of genes sampled without replacement in each iteration.}

\item{frac.col}{numeric. Fraction of samples sampled without replacement in each iteration.}

\item{n.cores}{numeric. The number of cores to use.}

\item{rank.ties}{a character string indicating how ties are treated in ranking. Accepted values include \code{"average", "first", "last", "random", "max", "min", "dense"}. For details, see \code{ties.method} parameter of the  \code{frank} function in  \code{data.table} package.}

\item{filter.mat}{matrix or NULL. Which edges to exclude.}

\item{weight.method}{a character string indicating how the weights are assigned. It must be either \code{"qnorm"} (default) or \code{"inverse.rank"}.}

\item{adjust.weight}{logical. Should the weights be raised to a power to exhibit a scale-free topology.}

\item{powerVector}{numeric vector. Powers for which the scale free topology fit indices are to be calculated.}

\item{RsquaredCut}{numeric. Desired minimum scale free topology fitting index R^2.}

\item{removeFirst}{logical. Should the first bin be removed from the connectivity histogram while computing scale-free topology fitting indices?}

\item{nBreaks}{numeric. Number of bins in connectivity histograms.}

\item{adjust.clr}{logical. Should the network weights be adjusted using context likelihood or relatedness? See \code{minet::clr} for details.}

\item{seed}{integer or NULL. Random number generator seed.}

\item{verbose}{logical or numeric. Print messages if verbose > 0.}
}
\value{
Returns a matrix representing a co-expression network.
}
\description{
This function reconstructs a co-expression network from gene expression profiles.
}
\details{
Note: higher values in \code{net} should correspond to higher confidence in the edge. If the sign of values mean positive or negative association between genes, you probably should provide absolute values.

This function does not yet handle directed networks.
}
\examples{
genes = c('TP53', 'RBM3', 'SF3', 'LIM12', 'ATM', 'TMEM160', 'BCL2L1', 'MDM2', 'PDR', 'MEG3', 'EGFR', 'CD96', 'KEAP1', 'SRSF1', 'TSEN2')
dummy_net = matrix(rnorm(length(genes)^2), nrow = length(genes), dimnames = list(genes, genes))
dummy_net = (dummy_net + t(dummy_net))/2 # symmetric undirected nework
cor_res = coexpression_ppi_cor(net = dummy_net)
print(sprintf('Pearson r: \%s, p: \%s', cor_res$pearson$estimate, cor_res$pearson$p.value))
print(sprintf('Spearman roh: \%s, p: \%s', cor_res$spearman$estimate, cor_res$spearman$p.value))
print(sprintf('Kendall tau: \%s, p: \%s', cor_res$kendall$estimate, cor_res$kendall$p.value))
}
